+++ RetroFE/Source/Graphics/Component/ReloadableText.cpp
@@ -93,7 +93,7 @@ bool ReloadableText::update(float dt)
     }
     else if (type_ == "file")
     {
-        Uint32 now = SDL_GetTicks();
+        Uint64 now = SDL_GetTicks();
         if (now - lastFileReloadTime_ >= fileDebounceDuration_) {
             ReloadTexture();
             lastFileReloadTime_ = now;
